fix: break up TitleBar into components
