+++ src/pages/view/TitleBar/styles/StarDashboardButton.module.css
@@ -1,0 +1,13 @@
+.star {
+    margin: 0 0 0 var(--spacers-dp4);
+    padding: 0;
+    border: none;
+    position: relative;
+    top: 1px;
+    background: transparent;
+    outline: none;
+    cursor: pointer;
+}
+.star:disabled {
+    cursor: not-allowed;
+}
